requirements.txt
numpy==1.17.0
Pillow==6.1.0
scikit-image==0.15.0
nevergrad==0.2.2
